+++ modules/app-desktop/src/main/kotlin/org/archivekeep/app/desktop/ui/components/ItemManySelect.kt
@@ -1,20 +1,15 @@
 package org.archivekeep.app.desktop.ui.components
 
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.ColumnScope
 import androidx.compose.foundation.layout.defaultMinSize
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.MutableState
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
+import org.archivekeep.app.desktop.ui.designsystem.dialog.LabelText
 import org.archivekeep.app.desktop.ui.designsystem.input.CheckboxWithText
 import org.archivekeep.app.desktop.ui.designsystem.input.TriStateCheckboxWithText
 
@@ -22,13 +17,43 @@ import org.archivekeep.app.desktop.ui.designsystem.input.TriStateCheckboxWithTex
 fun <I> ColumnScope.ItemManySelect(
     label: String,
     allItemsLabel: (count: Int) -> String,
-    itemLabel: (item: I) -> String,
+    itemLabelText: (item: I) -> String,
     allItems: List<I>,
     selectedItems: MutableState<Set<I>>,
 ) {
     val state = rememberManySelect(allItems, selectedItems)
 
-    Text(label, fontSize = 10.sp)
+    ItemManySelect(
+        label,
+        allItemsLabel,
+        itemLabelText = itemLabelText,
+        state = state,
+    )
+}
+
+@Composable
+fun <I, I_S> ColumnScope.ItemManySelect(
+    label: String,
+    allItemsLabel: (count: Int) -> String,
+    itemLabelText: (item: I) -> String,
+    state: ManySelectState<I, I, I_S>,
+) where I : I_S {
+    ItemManySelect(
+        label,
+        allItemsLabel,
+        itemLabel = { Text(itemLabelText(it)) },
+        state = state,
+    )
+}
+
+@Composable
+fun <I, I_S> ColumnScope.ItemManySelect(
+    label: String,
+    allItemsLabel: (count: Int) -> String,
+    itemLabel: @Composable (item: I) -> Unit,
+    state: ManySelectState<I, I, I_S>,
+) where I : I_S {
+    LabelText(label)
 
     TriStateCheckboxWithText(
         state.selectAllState,
@@ -40,27 +65,19 @@ fun <I> ColumnScope.ItemManySelect(
                 .defaultMinSize(minHeight = 32.dp),
     )
 
-    Box(
-        modifier =
-            Modifier
-                .background(Color(0xFFF9F9F9))
-                .verticalScroll(rememberScrollState())
-                .weight(weight = 1f, fill = false),
-    ) {
-        Column {
-            state.allItems.forEach { item ->
-                CheckboxWithText(
-                    state.selectedItems.contains(item),
-                    text = itemLabel(item),
-                    onValueChange = {
-                        state.onItemChange(item, it)
-                    },
-                    modifier =
-                        Modifier
-                            .fillMaxWidth()
-                            .defaultMinSize(minHeight = 24.dp),
-                )
-            }
+    Column {
+        state.allItems.forEach { item ->
+            CheckboxWithText(
+                state.selectedItems.contains(item),
+                text = { itemLabel(item) },
+                onValueChange = {
+                    state.onItemChange(item, it)
+                },
+                modifier =
+                    Modifier
+                        .fillMaxWidth()
+                        .defaultMinSize(minHeight = 24.dp),
+            )
         }
     }
 }