+++ modules/app-desktop/src/main/kotlin/org/archivekeep/app/desktop/ui/designsystem/dialog/LabelText.kt
@@ -2,9 +2,11 @@ package org.archivekeep.app.desktop.ui.designsystem.dialog
 
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.sp
 
 @Composable
 fun LabelText(text: String) {
-    Text(text, fontSize = 11.sp)
+    Text(text, fontSize = 11.sp, fontWeight = FontWeight.Bold, color = Color.DarkGray)
 }