+++ modules/app-desktop/src/main/kotlin/org/archivekeep/app/desktop/ui/dialogs/repository/operations/indexupdate/UpdateIndexOperationDialog.kt
@@ -164,7 +164,7 @@ class UpdateIndexOperationDialog(
                 ItemManySelect(
                     "Moves",
                     allItemsLabel = { "All moves ($it)" },
-                    itemLabel = { "${it.from} -> ${it.to}" },
+                    itemLabelText = { "${it.from} -> ${it.to}" },
                     allItems = state.preparedResult.moves,
                     selectedMoves,
                 )