+++ modules/app-desktop/src/main/kotlin/org/archivekeep/app/desktop/ui/dialogs/repository/operations/sync/RepoToRepoSyncUserFlow.kt
@@ -1,8 +1,9 @@
 package org.archivekeep.app.desktop.ui.dialogs.repository.operations.sync
 
+import androidx.compose.runtime.MutableState
+import androidx.compose.runtime.mutableStateOf
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.flatMapLatest
 import kotlinx.coroutines.flow.map
@@ -10,7 +11,8 @@ import kotlinx.coroutines.flow.stateIn
 import org.archivekeep.app.core.operations.sync.RepoToRepoSync
 import org.archivekeep.app.core.operations.sync.RepoToRepoSync.JobState
 import org.archivekeep.app.desktop.utils.stickToFirstNotNullAsState
-import org.archivekeep.files.operations.RelocationSyncMode
+import org.archivekeep.files.operations.sync.RelocationSyncMode
+import org.archivekeep.files.operations.sync.SyncSubOperation
 import org.archivekeep.utils.loading.Loadable
 import org.archivekeep.utils.loading.mapIfLoadedOrDefault
 import org.archivekeep.utils.loading.mapToLoadable
@@ -20,15 +22,12 @@ class RepoToRepoSyncUserFlow(
     val sync: RepoToRepoSync,
 ) {
     companion object {
-        val defaultRelocationSyncMode =
-            RelocationSyncMode.Move(
-                allowDuplicateIncrease = false,
-                allowDuplicateReduction = false,
-            )
+        val relocationSyncMode = RelocationSyncMode.Move(allowDuplicateIncrease = true, allowDuplicateReduction = true)
     }
 
     data class State(
         val operation: Loadable<RepoToRepoSync.State>,
+        val selectedOperations: MutableState<Set<SyncSubOperation>>,
     ) {
         val isRunning = operation.mapIfLoadedOrDefault(false) { it is JobState.Running }
         val isCancelled = operation.mapIfLoadedOrDefault(false) { it is JobState.Finished && it.cancelled }
@@ -40,22 +39,20 @@ class RepoToRepoSyncUserFlow(
 
     val currentOperation = sync.currentJobFlow.stickToFirstNotNullAsState(scope)
 
-    val relocationSyncModeFlow = MutableStateFlow<RelocationSyncMode>(defaultRelocationSyncMode)
+    val selectedOperations = mutableStateOf<Set<SyncSubOperation>>(emptySet())
 
     @OptIn(ExperimentalCoroutinesApi::class)
     val operationStateFlow =
         currentOperation
             .flatMapLatest {
                 if (it == null) {
-                    relocationSyncModeFlow.flatMapLatest { relocationSyncMode ->
-                        sync.prepare(relocationSyncMode)
-                    }
+                    sync.prepare(relocationSyncMode)
                 } else {
                     it.currentState.map { it as RepoToRepoSync.State }.mapToLoadable()
                 }
             }.stateIn(scope, SharingStarted.WhileSubscribed(), Loadable.Loading)
 
-    val stateFlow = operationStateFlow.map { State(operation = it) }
+    val stateFlow = operationStateFlow.map { State(operation = it, selectedOperations) }
 
     fun launch() {
         val stateLoadable =
@@ -64,7 +61,7 @@ class RepoToRepoSyncUserFlow(
 
         val state = stateLoadable.value as? RepoToRepoSync.State.Prepared ?: throw IllegalStateException("Must be in prepared state")
 
-        state.startExecution()
+        state.startExecution(selectedOperations.value)
     }
 
     fun cancel() {