+++ modules/files/src/main/kotlin/org/archivekeep/files/operations/sync/SyncSubOperation.kt
@@ -1,0 +1,11 @@
+package org.archivekeep.files.operations.sync
+
+import org.archivekeep.files.repo.Repo
+
+sealed interface SyncSubOperation {
+    suspend fun apply(
+        base: Repo,
+        dst: Repo,
+        logger: SyncLogger,
+    )
+}