+++ modules/files/src/main/kotlin/org/archivekeep/testing/fixtures/FixtureRepoBuilder.kt
@@ -1,8 +1,8 @@
 package org.archivekeep.testing.fixtures
 
 class FixtureRepoBuilder(
-    base: Map<String, String>,
-    baseUncommitted: Map<String, String>,
+    base: Map<String, String> = emptyMap(),
+    baseUncommitted: Map<String, String> = emptyMap(),
 ) {
     val repoContents = base.toMutableMap()
     val repoUncommittedContents = baseUncommitted.toMutableMap()
@@ -24,4 +24,11 @@ class FixtureRepoBuilder(
     fun deletePattern(regex: Regex) {
         repoContents.keys.removeAll(regex::matches)
     }
+
+    fun addFrom(fixtureRepo: FixtureRepo) {
+        repoContents.putAll(fixtureRepo.contents)
+        repoUncommittedContents.putAll(fixtureRepo.uncommittedContents)
+    }
+
+    fun build() = FixtureRepo(repoContents, repoUncommittedContents)
 }